+++ Calculus.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31410.357
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calculadora", "Calculadora\Calculadora.csproj", "{4EF421B0-0893-44FF-BAA1-386C5B30217B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4EF421B0-0893-44FF-BAA1-386C5B30217B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4EF421B0-0893-44FF-BAA1-386C5B30217B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4EF421B0-0893-44FF-BAA1-386C5B30217B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4EF421B0-0893-44FF-BAA1-386C5B30217B}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {6149F570-2CD3-49B3-8823-352B4E0A5040}
+	EndGlobalSection
+EndGlobal
