--- Calculus/Calculadora.csproj.user
@@ -1,9 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup />
-  <ItemGroup>
-    <Compile Update="frmCalculus.cs">
-      <SubType>Form</SubType>
-    </Compile>
-  </ItemGroup>
-</Project>
