--- Calculus/Resources/recursos.txt
@@ -1,1 +1,0 @@
-
